% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instload.packages.R
\name{instload.packages}
\alias{instload.packages}
\title{Install and Load R Packages}
\usage{
instload.packages(..., silent = FALSE)
}
\description{
This function receives a list of packages to install and load.
For each of the package names, it does:
      
      1- checks local repository of installed packages. If the given package is not installed
         it installs it otherwise it throws warning message and skip to the next package
      2- Once all packages installed successfully, it load them into workspace using 'require' function
If at all you give a package that cannot be installed, it ignores its loading and throws a warning message
about its failure to install the package.
}
\details{
Note 1: Functional Modes: By default the function reports successfull installation of those packages successfully installed.
However, if reporting is not desirable, can turn it off by 'silent=FALSE'

Note 2: instload.packages does accept only list of characters. It does not accept integer or decimal values

instload.packages("pkg1","pkg2",....,"pkgn")

example:
instload.packages("caret","tm","ts")
instload.packages("caret","tm","ts",silent=TRUE)
==============================================
Please do not hesiatate to report errors or exceptions to my email @ Abolfazli.s@gmail.com
}

